Move EditorLauncher to library to fix errors with enum attributes.
